pyproject.toml
[project]
name = "monorepo-python-template"
version = "0.1.0"
description = "Monorepo Python Template"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry]
name = "monorepo-python-template"
version = "0.1.0"
description = "Monorepo Python Template"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-cov = "^4.1.0"
black = "^23.3.0"
ruff = "^0.1.7"
mypy = "^1.0.0"
pydoclint = "^0.6.0"
pymarkdownlnt = "^0.9.0"
yamllint = "^1.32.0"
pre-commit = "^4.2.0"

# Subproject dependency groups will be added by create_python_subproject.sh

[tool.black]
line-length = 120
target-version = ['py310']
exclude = '''
/(
    \.git
    | \.venv
)/
'''
extend-exclude = ".*/*.ipynb|.*/*.xls"

[tool.ruff]
lint.select = ["E", "F", "B", "D", "I"]
lint.ignore = ["D104", "D107", "D203", "D205", "D212", "D401", "D406", "D407"]
lint.fixable = ["E", "F", "B", "I"]
line-length = 120
target-version = "py310"
exclude = [
    "__pycache__",
    ".ruff_cache",
    ".mypy_cache",
    ".pytest_cache",
    "*.log",
    ".venv",
]

[tool.pydoclint]
style = "google"
check-return = true
check-arg-order = true
exclude = "(__pycache__|\\.ruff_cache|\\.mypy_cache|\\.pytest_cache|\\.venv|.*\\.log)"

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
disable_error_code = ["name-defined", "arg-type", "dict-item"]
explicit_package_bases = true
exclude = "(__pycache__|\\.ruff_cache|\\.mypy_cache|\\.pytest_cache|\\.venv|.*\\.log)"

[tool.pymarkdown]
plugins.MD013.enabled = false
plugins.MD024.enabled = false
plugins.MD036.enabled = false
plugins.MD037.enabled = false
plugins.MD033 = { allowed_elements = ["br"] }

[tool.yamllint]
extends = "default"
ignore = [
    "__pycache__",
    ".ruff_cache",
    ".pytest_cache",
    ".mypy_cache",
    "*.log",
    ".venv",
]

[tool.yamllint.rules]
line-length.disable = true
document-start.disable = true
comments-indentation.disable = true
truthy.disable = true

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "*/__pycache__/*",
    "*/tests/*",
    "*/.venv/*",
    "*/.ruff_cache/*",
    "*/.mypy_cache/*",
    "*/.pytest_cache/*",
    "*.log"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"






[tool.poetry.group.core.dependencies]
azure-monitor-opentelemetry = "1.6.5"
python-dotenv = "1.0.1"
azure-identity = "1.20.0"
nest-asyncio = "1.6.0"
opentelemetry-api = "1.31.1"
semantic-kernel = "1.27.1"
openai = "^1.65.5"
azure-core = "^1.32.0"
azure-cosmos = "^4.9.0"
azure-storage-blob = "^12.16.0"
pydantic = "^2.10.6"
streamlit = "^1.44.1"
authlib = "^1.3.2"
plotly = "^6.0.1"
orjson = "^3.10.15"
llama-index = "^0.12.30"
llama-index-embeddings-azure-openai = "^0.3.2"
llama-index-llms-azure-openai = "^0.3.2"
llama-index-embeddings-openai = "^0.3.1"
tabulate = "^0.9.0"
opentelemetry-instrumentation-tornado = "0.52b1"






[tool.poetry.group.core]
optional = true
